feat : Port and PersistenceAdapter refactoring
