+++ src/main/java/com/econovation/idp/adapter/in/web/AccountController.java
@@ -1,12 +1,12 @@
 package com.econovation.idp.adapter.in.web;
 
-import com.econovation.idp.application.port.in.LoginRequestDto;
-import com.econovation.idp.application.port.in.SignUpRequestDto;
-import com.econovation.idp.application.port.out.LoginResponseDto;
+import com.econovation.idp.domain.dto.LoginRequestDto;
+import com.econovation.idp.domain.dto.SignUpRequestDto;
+import com.econovation.idp.application.port.in.JwtProviderUseCase;
+import com.econovation.idp.domain.dto.LoginResponseDto;
 
 import javax.validation.Valid;
-import com.econovation.idp.application.service.AccountJwtService;
-import com.econovation.idp.global.config.jwt.JwtProvider;
+import com.econovation.idp.application.port.in.AccountJwtUseCase;
 import com.econovation.idp.application.service.AccountSignUpService;
 import com.econovation.idp.global.common.BasicResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -15,12 +15,10 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PostAuthorize;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -39,12 +37,9 @@ import java.util.Map;
 @CrossOrigin(origins = "*")
 @Tag(name = "Account 관련 서비스", description = "회원가입, 로그인 등등")
 public class AccountController {
-    private final AccountJwtService accountJwtService;
+    private final AccountJwtUseCase accountJwtUseCase;
     private final AccountSignUpService accountSignUpService;
-    private final JwtProvider jwtProvider;
-
-    @Value("${login.redirect_url}")
-    private String loginPageUrl;
+    private final JwtProviderUseCase jwtProviderUseCase;
 
     //    로그아웃 기능 구현
     @Operation(summary = "logout", description = "로그아웃_에이전트, 로그아웃시 redirect 페이지로 이동")
@@ -54,7 +49,7 @@ public class AccountController {
 //        7번부터 빼야 bearer(+스페이스바) 빼고 토큰만 추출 가능
         String refreshToken = request.getHeader("Authorization").substring(7);
 
-        accountJwtService.logout(refreshToken);
+        accountJwtUseCase.logout(refreshToken);
         URI redirectUri = new URI(redirectUrl);
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setLocation(redirectUri);
@@ -70,10 +65,10 @@ public class AccountController {
     })
     @GetMapping("/api/account/re-issue")
     public ResponseEntity<LoginResponseDto> reIssue(@Email String userEmail, String refreshToken, HttpServletRequest request) {
-        if(!jwtProvider.validateToken(request,refreshToken).isAuthenticated()){
+        if(!jwtProviderUseCase.validateToken(request,refreshToken).isAuthenticated()){
             return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
-        LoginResponseDto responseDto = accountJwtService.reIssueAccessToken(userEmail, refreshToken);
+        LoginResponseDto responseDto = accountJwtUseCase.reIssueAccessToken(userEmail, refreshToken);
         return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
 
@@ -109,16 +104,16 @@ public class AccountController {
             @ApiResponse(responseCode = "HttpStatus.OK", description = "로그인 내부 인증 처리")
     })
     @PostMapping("/api/account/login/process")
-    public ResponseEntity<LoginResponseDto> login(@Valid LoginRequestDto loginDto) throws URISyntaxException {
-        LoginResponseDto responseDto = accountJwtService.login(loginDto.getUserEmail(), loginDto.getPassword());
+    public ResponseEntity<Map> login(@Valid LoginRequestDto loginDto) throws URISyntaxException {
+        LoginResponseDto responseDto = accountJwtUseCase.login(loginDto.getUserEmail(), loginDto.getPassword());
         URI redirectUri = new URI(loginDto.getRedirectUrl());
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setLocation(redirectUri);
         log.info("redirectUrl" + redirectUri);
-        Date expiredTime = jwtProvider.getExpiredTime(responseDto.getRefreshToken());
+        Date expiredTime = jwtProviderUseCase.getExpiredTime(responseDto.getRefreshToken());
         Map<Date, LoginResponseDto> loginResponseDto = new HashMap<>();
         loginResponseDto.put(expiredTime, responseDto);
-        return new ResponseEntity<>(responseDto, httpHeaders, HttpStatus.OK);
+        return new ResponseEntity<>(loginResponseDto, httpHeaders, HttpStatus.OK);
     }
 
     @PostAuthorize("hasRole('ROLE_USER')")
@@ -129,7 +124,7 @@ public class AccountController {
     })
     @GetMapping("/api/account/re-check")
     public ResponseEntity<BasicResponse> checkValideToken(HttpServletRequest request, String refreshToken) {
-        Authentication authentication = jwtProvider.validateToken(request, refreshToken);
+        Authentication authentication = jwtProviderUseCase.validateToken(request, refreshToken);
         if (!authentication.isAuthenticated()) {
 
             return new ResponseEntity<>(HttpStatus.OK);
@@ -146,8 +141,8 @@ public class AccountController {
     })
     @PostMapping("/api/account/login/process/expired")
     public ResponseEntity<Map<Date,LoginResponseDto>> loginWithExpiredTime(@Valid LoginRequestDto loginDto) throws URISyntaxException {
-        LoginResponseDto responseDto = accountJwtService.login(loginDto.getUserEmail(), loginDto.getPassword());
-        Date expiredTime = jwtProvider.getExpiredTime(responseDto.getRefreshToken());
+        LoginResponseDto responseDto = accountJwtUseCase.login(loginDto.getUserEmail(), loginDto.getPassword());
+        Date expiredTime = jwtProviderUseCase.getExpiredTime(responseDto.getRefreshToken());
         URI redirectUri = new URI(loginDto.getRedirectUrl());
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setLocation(redirectUri);
