--- src/main/java/com/econovation/idp/application/mapper/Mapper.java
@@ -1,4 +1,0 @@
-package com.econovation.idp.application.mapper;
-
-public class Mapper {
-}
