+++ src/main/java/com/econovation/idp/application/port/in/AccountJwtUseCase.java
@@ -1,0 +1,10 @@
+package com.econovation.idp.application.port.in;
+
+import com.econovation.idp.domain.dto.LoginResponseDto;
+
+public interface AccountJwtUseCase {
+    public LoginResponseDto
+    reIssueAccessToken(String email, String refreshedToken);
+    public LoginResponseDto login(String email, String password);
+    public void logout(String refreshToken);
+}
