+++ src/main/java/com/econovation/idp/application/port/in/AccountUseCase.java
@@ -1,4 +1,5 @@
 package com.econovation.idp.application.port.in;
 
 public interface AccountUseCase {
+
 }
