--- src/main/java/com/econovation/idp/application/port/in/LoginRequestDto.java
@@ -1,10 +1,0 @@
-package com.econovation.idp.application.port.in;
-
-import lombok.Getter;
-
-@Getter
-public class LoginRequestDto {
-    private String userEmail;
-    private String password;
-    private String redirectUrl;
-}
