+++ src/main/java/com/econovation/idp/application/port/in/UserUseCase.java
@@ -1,5 +1,7 @@
 package com.econovation.idp.application.port.in;
 
+import com.econovation.idp.domain.dto.UserPasswordUpdateDto;
+import com.econovation.idp.domain.dto.UserUpdateRequestDto;
 import com.econovation.idp.domain.user.Account;
 
 import java.util.List;
@@ -13,7 +15,7 @@ public interface UserUseCase {
     List<Account> findUserByRole(int page, String role);
     Account indUserByYearAndUserName(String userName,Long year);
     Account findUserByUserEmail(String userEmail);
-    Account updateUser(Long userId,UserUpdateRequestDto userUpdateRequestDto);
+    Account updateUser(Long userId, UserUpdateRequestDto userUpdateRequestDto);
     void deleteUserById(Long userId);
     Account setPassword(UserPasswordUpdateDto userPasswordUpdateDto);
 }
