+++ src/main/java/com/econovation/idp/domain/dto/SignUpRequestDto.java
@@ -1,22 +1,20 @@
-package com.econovation.idp.application.port.in;
+package com.econovation.idp.domain.dto;
 
+import com.econovation.idp.domain.auth.Password;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 
 @Getter
 @AllArgsConstructor
 public class SignUpRequestDto {
-    @NotNull
+    @Email
     private String userEmail;
-    @NotNull
     private Long year;
-    @NotNull
     private String userName;
-    @NotNull
+    @Password
     private String password;
-    @NotNull
-    private String pinCode;
 }
