+++ src/main/java/com/econovation/idp/domain/dto/UserFindDto.java
@@ -1,29 +1,25 @@
-package com.econovation.idp.application.port.in;
+package com.econovation.idp.domain.dto;
 
 import com.econovation.idp.domain.user.Account;
 import lombok.AccessLevel;
+import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Range;
 
-import javax.validation.constraints.NotNull;
-
+@Data
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserPasswordUpdateDto {
+public class UserFindDto {
     @Range(min =1, max = 50)
     private Long year;
-    @NotNull
-    private String userName;
 
-    private String password;
+    private String userName;
 
     public Account toEntity(){
         return Account.builder()
                 .year(year)
-                .userName(userName)
-                .password(password).
+                .userName(userName).
                 build();
     }
 }
-
