+++ src/main/java/com/econovation/idp/domain/dto/UserPasswordUpdateDto.java
@@ -1,25 +1,29 @@
-package com.econovation.idp.application.port.in;
+package com.econovation.idp.domain.dto;
 
 import com.econovation.idp.domain.user.Account;
 import lombok.AccessLevel;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Range;
 
-@Data
+import javax.validation.constraints.NotNull;
+
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserFindDto {
+public class UserPasswordUpdateDto {
     @Range(min =1, max = 50)
     private Long year;
-
+    @NotNull
     private String userName;
+
+    private String password;
 
     public Account toEntity(){
         return Account.builder()
                 .year(year)
-                .userName(userName).
+                .userName(userName)
+                .password(password).
                 build();
     }
 }
+
