+++ src/main/java/com/econovation/idp/global/config/SecurityConfig.java
@@ -1,32 +1,20 @@
 package com.econovation.idp.global.config;
 
-import com.econovation.idp.application.service.AccountJwtService;
-import com.econovation.idp.application.service.AccountJwtServiceImpl;
-import com.econovation.idp.application.service.AccountSignUpService;
+import com.econovation.idp.application.port.in.AccountJwtUseCase;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.econovation.idp.global.common.BasicResponse;
 import com.econovation.idp.global.config.jwt.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.actuate.autoconfigure.security.servlet.EndpointRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.config.web.server.ServerHttpSecurity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.server.SecurityWebFilterChain;
 
 import java.io.PrintWriter;
-
-import static org.springframework.security.config.Customizer.withDefaults;
 
 @EnableWebSecurity
 @RequiredArgsConstructor
@@ -34,7 +22,7 @@ public class SecurityConfig {
     private final JwtAuthenticationFilter jwtAuthenticationFilter;
     private final ObjectMapper objectMapper;
 
-    private AccountJwtService accountJwtService;
+    private AccountJwtUseCase accountJwtUseCase;
     private final String UNAUTHORIZEd_CUSTOM_MESSAGE = "인증받지 못한 유저입니다. 로그인을 재시도해주세요.";
 
     @Bean
